perf/intel: Remove Perfmon-v4 counter_freezing support
mainline inclusion
from mainline-v5.12-rc1
commit 3daa96d67274653b7c461b30ef9581d68e905fe1
category: feature
feature: SPR PMU core event enhancement
bugzilla: https://gitee.com/openeuler/intel-kernel/issues/I596BF

Intel-SIG: commit 3daa96d67274 perf/intel: Remove Perfmon-v4
counter_freezing support
This commit is backported as a dependency for SPR core PMU support.

--------------------------------

Perfmon-v4 counter freezing is fundamentally broken; remove this default
disabled code to make sure nobody uses it.

The feature is called Freeze-on-PMI in the SDM, and if it would do that,
there wouldn't actually be a problem, *however* it does something subtly
different. It globally disables the whole PMU when it raises the PMI,
not when the PMI hits.

This means there's a window between the PMI getting raised and the PMI
actually getting served where we loose events and this violates the
perf counter independence. That is, a counting event should not result
in a different event count when there is a sampling event co-scheduled.

This is known to break existing software (RR).

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Yunying Sun <yunying.sun@intel.com>
Signed-off-by: Jun Tian <jun.j.tian@intel.com>